+++ RELEASE_NOTES.rst
@@ -9,7 +9,8 @@ Unreleased
 Highlights:
 
 * Implemented Phaser-servo. This requires recent gateware on Phaser.
-
+* Implemented Phaser-MIQRO support. This requires the Phaser MIQRO gateware
+  variant.
 
 ARTIQ-7
 -------